simplify schema create form
