+++ playground/src/components/SchemaPatchForm.tsx
@@ -1,8 +1,13 @@
 import { useState } from 'react'
+import { useForm, useFieldArray } from 'react-hook-form'
 import { useSchemaContext } from '@graphiql/react'
 import { useQueryClient } from '@tanstack/react-query'
-import { SchemaForm, FormData } from './SchemaForm'
-import { patchSchema, ErrorItem } from '../lib/api'
+import { patchSchema, Field, ErrorItem } from '../lib/api'
+
+export type FormData = {
+  name: string
+  fields: Field[]
+}
 
 export type SchemaPatchFormProps = {
   values?: FormData
@@ -15,6 +20,9 @@ export function SchemaPatchForm({ values, fieldTypes }: SchemaPatchFormProps) {
 
   const [errors, setErrors] = useState<ErrorItem[]>()
   const [isLoading, setIsLoading] = useState(false)
+
+  const { control, register, handleSubmit } = useForm<FormData>({ values })
+  const { fields, append, remove } = useFieldArray({ control, name: 'fields', keyName: '_id' })
 
   const onSubmit = async (data: FormData) => {
     setErrors(undefined)
@@ -36,12 +44,64 @@ export function SchemaPatchForm({ values, fieldTypes }: SchemaPatchFormProps) {
   }
 
   return (
-    <SchemaForm
-      errors={errors}
-      isLoading={isLoading}
-      onSubmit={onSubmit}
-      values={values}
-      fieldTypes={fieldTypes}
-    />
+    <form 
+      className="graphiql-defradb-form"
+      onSubmit={handleSubmit(onSubmit)}
+    >
+      {errors?.map((error, index) =>
+        <div key={index} className="graphiql-defradb-error">
+          <span>{error.message}</span>
+        </div>
+      )}
+      <div className="graphiql-defradb-field-header">
+        <h5 className="graphiql-defradb-input-label">Fields</h5>
+        <button
+          type="button"
+          className="graphiql-button"
+          onClick={() => append({ name: '', kind: 'String', internal: false })}
+        >
+          Add
+        </button>
+      </div>
+      {fields.map((field, index) =>
+        <div 
+          key={field._id} 
+          className="graphiql-defradb-field" 
+          style={{ display: field.internal ? 'none' : undefined }}
+        >
+          <input
+            className="graphiql-defradb-input"
+            disabled={isLoading || !!field.id}
+            {...register(`fields.${index}.name`)}
+          />
+          <select
+            className="graphiql-defradb-input"
+            disabled={isLoading || !!field.id}
+            {...register(`fields.${index}.kind`)}
+          >
+            {fieldTypes.map((value, index) => 
+              <option key={index} value={value}>{value}</option>
+            )}
+          </select>
+          {!field.id &&
+            <button
+              type="button"
+              className="graphiql-button"
+              onClick={() => remove(index)}
+              disabled={isLoading || !!field.id}
+            >
+              Remove
+            </button>
+          }
+        </div>
+      )}
+      <button 
+        type="submit"
+        className="graphiql-button"
+        disabled={isLoading}
+      >
+        Submit
+      </button>
+    </form>
   ) 
 }
